+++ src/variety/tents.js
@@ -259,10 +259,10 @@
 
 	BoardExec: {
 		adjustBoardData: function(key, d) {
-			this.adjustEdge_1(key, d);
+			this.adjustExCellTopLeft_1(key, d);
 		},
 		adjustBoardData2: function(key, d) {
-			this.adjustEdge_2(key, d);
+			this.adjustExCellTopLeft_2(key, d);
 		}
 	},