+++ Changelog.md
@@ -2,6 +2,7 @@
 
 * araf: New puzzle type Araf (due to Giuliano Montelucci)
 * balance: New puzzle type Balance Loop (due to Giuliano Montelucci)
+* doppelblock: New puzzle type Doppelblock
 * maxi: New puzzle type Maxi Loop (due to Giuliano Montelucci)
 * midloop: New puzzle type Mid-loop (due to Lennard Sprong)
 * simpleloop: New puzzle type Simple Loop
@@ -25,6 +26,7 @@
 * Remove several options.
 * Allow specifying that a puzzle is a variant.
 * Don't start timer until puzzle is visible.
+* Center grid dashing.
 
 #### Bugfixes